chore :: 베포하면서 필요한 설정 변경

+++ database/migrations/2025_06_30_012304_add_courier_to_insert_table.php
@@ -1,19 +1,21 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
-return new class extends Migration
+class AddCourierToInsertTable extends Migration
 {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('insert', function (Blueprint $table) {
-            $table->string('courier')->nullable()->after('tracking_number');
-        });
+        if (!Schema::hasColumn('insert', 'courier')) {
+            Schema::table('insert', function (Blueprint $table) {
+                $table->string('courier')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('insert', function (Blueprint $table) {
-            $table->dropColumn('courier');
-        });
+        if (Schema::hasColumn('insert', 'courier')) {
+            Schema::table('insert', function (Blueprint $table) {
+                $table->dropColumn('courier');
+            });
+        }
     }
-};
+}
